ATS Review Using Gemini AI/requirements.txt
langchain
streamlit
google-generativeai
python-dotenv
pdf2image